Add files via upload
This analysis is done on a dataset collected by a health insurance company in a survey of 3,81,109 policyholders of the company. The company wishes to enter the vehicle insurance market. For a start, the company wants to analyse the factors that guide  customer's decision to buy the vehicle insurance policy proposed by the company. The dataset has information about the gender, age, and region code of the survey participants, as well as information about whether they have previously taken any insurance policy, age of their vehicle, whether vehicle has damage, the annual premium they pay, vintage and finally, their response to the proposed insurance policy of the company.